Abstract Sync page logic into a provider pattern.
