+++ assets/js/sync-ui/apps/settings-page.js
@@ -7,34 +7,82 @@ import { __, sprintf } from '@wordpress/i18n';
 /**
  * Internal dependencies.
  */
-import SyncPanel from './sync/panel';
+import { useSync } from '../../sync';
+import SyncPanel from '../components/sync/panel';
 
 /**
  * Sync page component.
  *
- * @param {object} props Component props.
- * @param {boolean} props.isCli If sync is a CLI sync.
- * @param {boolean} props.isComplete If sync is complete.
- * @param {boolean} props.isDeleting If sync is a delete and sync.
- * @param {boolean} props.isEpio If ElasticPress is using ElasticPress.io.
- * @param {boolean} props.isFailed If sync has failed.
- * @param {boolean} props.isPaused If sync is paused.
- * @param {boolean} props.isSyncing If sync is running.
- * @param {number} props.itemsProcessed Number of items processed.
- * @param {number} props.itemsTotal Number of items to process.
- * @param {string} props.lastSyncDateTime Date and time of last sync in ISO-8601.
- * @param {boolean} props.lastSyncFailed If the last sync had failures.
- * @param {object[]} props.log Sync message log.
- * @param {Function} props.onDelete Callback for clicking delete and sync.
- * @param {Function} props.onPause Callback for clicking pause.
- * @param {Function} props.onResume Callback for clicking resume.
- * @param {Function} props.onStop Callback for clicking stop.
- * @param {Function} props.onSync Callback for clicking sync.
- * @param {string} props.syncStartDateTime Date and time of current sync in ISO 8601.
  * @returns {WPElement} Sync page component.
  */
-export default ({ isCli, isComplete, isDeleting, isEpio, isFailed, isSyncing, log, ...props }) => {
-	const isInitialSync = props.lastSyncDateTime === null;
+export default () => {
+	const {
+		isCli,
+		isComplete,
+		isDeleting,
+		isEpio,
+		isFailed,
+		isSyncing,
+		lastSyncDateTime,
+		log,
+		logMessage,
+		pauseSync,
+		resumeSync,
+		startSync,
+		stopSync,
+	} = useSync();
+
+	const isInitialSync = lastSyncDateTime === null;
+
+	/**
+	 * Handle clicking delete and sync button.
+	 *
+	 * @returns {void}
+	 */
+	const onDelete = async () => {
+		startSync(true);
+		logMessage(__('Starting delete and sync…', 'elasticpress'), 'info');
+	};
+
+	/**
+	 * Handle clicking pause button.
+	 *
+	 * @returns {void}
+	 */
+	const onPause = () => {
+		pauseSync();
+		logMessage(__('Pausing sync…', 'elasticpress'), 'info');
+	};
+
+	/**
+	 * Handle clicking play button.
+	 *
+	 * @returns {void}
+	 */
+	const onResume = () => {
+		resumeSync();
+		logMessage(__('Resuming sync…', 'elasticpress'), 'info');
+	};
+
+	/**
+	 * Handle clicking stop button.
+	 *
+	 * @returns {void}
+	 */
+	const onStop = () => {
+		stopSync();
+		logMessage(__('Sync stopped', 'elasticpress'), 'info');
+	};
+
+	/**
+	 * Handle clicking sync button.
+	 *
+	 * @returns {void}
+	 */
+	const onSync = async () => {
+		startSync(false);
+		logMessage(__('Starting sync…', 'elasticpress'), 'info');
+	};
 
 	return (
 		<>
@@ -58,15 +106,17 @@ export default ({ isCli, isComplete, isDeleting, isEpio, isFailed, isSyncing, lo
 								'elasticpress',
 						  )
 				}
-				isComplete={isComplete}
 				isDisabled={(isSyncing && isDeleting) || (isSyncing && isCli)}
-				isFailed={isFailed}
-				isSyncing={isSyncing && !isDeleting}
+				isLoading={isSyncing && !isDeleting}
 				logMessages={log.filter((m) => !m.isDeleting)}
 				showLastSync
 				showProgress={!isDeleting && (isSyncing || isComplete || isFailed)}
 				showSync
-				{...props}
+				onDelete={onDelete}
+				onPause={onPause}
+				onResume={onResume}
+				onSync={onSync}
+				onStop={onStop}
 			/>
 
 			{!isInitialSync ? (
@@ -76,10 +126,8 @@ export default ({ isCli, isComplete, isDeleting, isEpio, isFailed, isSyncing, lo
 						'If you are still having issues with your search results, you may need to do a completely fresh sync.',
 						'elasticpress',
 					)}
-					isComplete={isComplete}
 					isDisabled={(isSyncing && !isDeleting) || (isSyncing && isCli)}
-					isFailed={isFailed}
-					isSyncing={isSyncing && isDeleting}
+					isLoading={isSyncing && isDeleting}
 					logMessages={log.filter((m) => m.isDeleting)}
 					showDelete
 					showProgress={isDeleting && (isSyncing || isComplete || isFailed)}
@@ -87,7 +135,11 @@ export default ({ isCli, isComplete, isDeleting, isEpio, isFailed, isSyncing, lo
 						'All indexed data on ElasticPress will be deleted without affecting anything on your WordPress website. This may take a few hours depending on the amount of content that needs to be synced and indexed. While this is happening, searches will use the default WordPress results',
 						'elasticpress',
 					)}
-					{...props}
+					onDelete={onDelete}
+					onPause={onPause}
+					onResume={onResume}
+					onSync={onSync}
+					onStop={onStop}
 				/>
 			) : null}
 		</>
