+++ assets/js/sync/src/utilities.js
